renaming to wintappy, changing sql file to lowercase
