+++ .gitignore
@@ -59,4 +59,4 @@ ENV/
 env.bak/
 venv.bak/
 
-wintap/examples/*.db
+wintappy/examples/*.db
