+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = wintap
+name = wintappy
 
 [options]
 packages = find:
@@ -16,7 +16,7 @@ datautils =
 
 [options.entry_points]
 console_scripts = 
-    collectbinaries = wintap.etlutils.collectbinaries:main
-    download_from_s3 = wintap.etlutils.download_from_s3:main
-    rawtorolling = wintap.etlutils.rawtorolling:main
-    rawtostdview = wintap.etlutils.rawtostdview:main
+    collectbinaries = wintappy.etlutils.collectbinaries:main
+    download_from_s3 = wintappy.etlutils.download_from_s3:main
+    rawtorolling = wintappy.etlutils.rawtorolling:main
+    rawtostdview = wintappy.etlutils.rawtostdview:main
