+++ wintappy/datautils/rawtostdview.sql
