+++ wintappy/datautils/rawutil.py
