+++ wintappy/datautils/stdview_duckdb.py
