+++ wintappy/datautils/stdviewutil.py
