+++ wintappy/etlutils/collectbinaries.py
