+++ wintappy/etlutils/rawtorolling.py
@@ -6,7 +6,7 @@ from importlib_resources import files
 
 from jinjasql import JinjaSql
 
-from wintap.datautils import rawutil as ru
+from wintappy.datautils import rawutil as ru
 
 
 def daterange(start_date, end_date):
@@ -21,7 +21,7 @@ def process_range(cur_dataset, start_date, end_date):
         globs = ru.get_globs_for(cur_dataset, daypk)
         # No need to pass dayPK as the globs already include it.
         ru.create_raw_views(con, globs)
-        ru.run_sql_no_args(con, files("wintap.datautils").joinpath("RawToStdView.sql"))
+        ru.run_sql_no_args(con, files("wintap.datautils").joinpath("rawtostdview.sql"))
         ru.write_parquet(
             con, cur_dataset, ru.get_db_objects(con, exclude=["tmp"]), daypk
         )
